Refactor BigqueryResult.Row to BigqueryRow
(First part of a larger change in how BigQuery data is exposed.)
