+++ apis/Google.Bigquery.V2/Google.Bigquery.V2/BigqueryResult.cs
@@ -57,14 +57,14 @@ namespace Google.Bigquery.V2
         /// The rows of the query. If there are more rows than were returned in the first page of results,
         /// iterating over this will lazily evaluate the results each time, making further requests as necessary.
         /// </summary>
-        public IEnumerable<Row> Rows { get; }
+        public IEnumerable<BigqueryRow> Rows { get; }
 
         internal BigqueryResult(BigqueryClient client, GetQueryResultsResponse firstResult, Func<GetQueryResultsRequest> requestProvider)
         {
             _client = client;
             Schema = firstResult.Schema;
             var firstRows = firstResult.Rows ?? new TableRow[0];
-            Rows = firstRows.Concat(GetRemainingRows(requestProvider, firstResult.PageToken)).Select(row => new Row(this, row));
+            Rows = firstRows.Concat(GetRemainingRows(requestProvider, firstResult.PageToken)).Select(row => new BigqueryRow(row, Schema));
         }
 
         internal BigqueryResult(BigqueryClient client, TableDataList firstResult, TableSchema schema, Func<TabledataResource.ListRequest> requestProvider)
@@ -72,7 +72,7 @@ namespace Google.Bigquery.V2
             _client = client;
             Schema = schema;
             var firstRows = firstResult.Rows ?? new TableRow[0];
-            Rows = firstRows.Concat(GetRemainingRows(requestProvider, firstResult.PageToken)).Select(row => new Row(this, row));
+            Rows = firstRows.Concat(GetRemainingRows(requestProvider, firstResult.PageToken)).Select(row => new BigqueryRow(row, Schema));
         }
 
         internal BigqueryResult(BigqueryClient client, QueryResponse syncResponse)
@@ -82,7 +82,7 @@ namespace Google.Bigquery.V2
             var reference = syncResponse.JobReference;
             Func<GetQueryResultsRequest> requestProvider = () => client.Service.Jobs.GetQueryResults(reference.ProjectId, reference.JobId);
             var firstRows = syncResponse.Rows ?? new TableRow[0];
-            Rows = firstRows.Concat(GetRemainingRows(requestProvider, syncResponse.PageToken)).Select(row => new Row(this, row));
+            Rows = firstRows.Concat(GetRemainingRows(requestProvider, syncResponse.PageToken)).Select(row => new BigqueryRow(row, Schema));
         }
 
         private IEnumerable<TableRow> GetRemainingRows(Func<GetQueryResultsRequest> requestProvider, string pageToken)
@@ -129,143 +129,6 @@ namespace Google.Bigquery.V2
                 }
             }
             throw new KeyNotFoundException($"No such field: '{fieldName}'");
-        }
-
-        /// <summary>
-        /// A row within the result.
-        /// </summary>
-        public sealed class Row
-        {
-            private readonly BigqueryResult _parent;
-            public TableRow RawRow { get; }
-
-            internal Row(BigqueryResult parent, TableRow rawRow)
-            {
-                _parent = parent;
-                RawRow = rawRow;
-            }
-
-            private static readonly Func<string, string> StringConverter = v => v;
-            private static readonly Func<string, long> Int64Converter = v => long.Parse(v, CultureInfo.InvariantCulture);
-            private static readonly Func<string, double> DoubleConverter = v => double.Parse(v, CultureInfo.InvariantCulture);
-            private static readonly Func<string, DateTime> DateTimeConverter = v => new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc).AddSeconds(DoubleConverter(v));
-            private static readonly Func<string, byte[]> BytesConverter = v => Convert.FromBase64String(v);
-            private static readonly Func<string, bool> BooleanConverter = v => v == "true";
-
-
-            /// <summary>
-            /// Retrieves a cell value by field name.
-            /// </summary>
-            public object this[string name] => this[_parent.GetFieldIndex(name)];
-
-            /// <summary>
-            /// Retrieves a cell value by index.
-            /// </summary>
-            public object this[int index]
-            {
-                get
-                {
-                    object rawValue = RawRow.F[index].V;
-                    var field = _parent.Schema.Fields[index];
-
-                    return ConvertSingleValue(rawValue, field);
-                }
-            }
-
-            private static object ConvertSingleValue(object rawValue, TableFieldSchema field)
-            {
-                if (rawValue == null)
-                {
-                    return null;
-                }
-                var type = field.GetFieldType();
-
-                if (field.GetFieldMode() == FieldMode.Repeated)
-                {
-                    JArray array = (JArray)rawValue;
-                    switch (type)
-                    {
-                        case BigqueryDbType.String:
-                            return ConvertArray(array, StringConverter);
-                        case BigqueryDbType.Integer:
-                            return ConvertArray(array, Int64Converter);
-                        case BigqueryDbType.Float:
-                            return ConvertArray(array, DoubleConverter);
-                        case BigqueryDbType.Bytes:
-                            return ConvertArray(array, BytesConverter);
-                        case BigqueryDbType.Boolean:
-                            return ConvertArray(array, BooleanConverter);
-                        case BigqueryDbType.Timestamp:
-                            return ConvertArray(array, DateTimeConverter);
-                        case BigqueryDbType.Record:
-                            return ConvertRecordArray(array, field);
-                        default:
-                            throw new InvalidOperationException($"Unhandled field type {type} {rawValue.GetType()}");
-                    }
-                }
-                switch (type)
-                {
-                    case BigqueryDbType.String:
-                        return StringConverter((string)rawValue);
-                    case BigqueryDbType.Integer:
-                        return Int64Converter((string)rawValue);
-                    case BigqueryDbType.Float:
-                        return DoubleConverter((string)rawValue);
-                    case BigqueryDbType.Bytes:
-                        return BytesConverter((string)rawValue);
-                    case BigqueryDbType.Boolean:
-                        return BooleanConverter((string)rawValue);
-                    case BigqueryDbType.Timestamp:
-                        return DateTimeConverter((string)rawValue);
-                    case BigqueryDbType.Record:
-                        return ConvertRecord((JObject)rawValue, field);
-                    default:
-                        throw new InvalidOperationException($"Unhandled field type {type} {rawValue.GetType()}");
-                }
-            }
-
-            // TODO: GetString etc, like IDataReader etc. (Should we actually implement IDataReader?)
-
-            private static T[] ConvertArray<T>(JArray array, Func<string, T> converter)
-            {
-                T[] ret = new T[array.Count];
-                for (int i = 0; i < ret.Length; i++)
-                {
-                    JObject value = (JObject)array[i];
-                    ret[i] = converter((string) value["v"]);
-                }
-                return ret;
-            }
-
-            private static Dictionary<string, object>[] ConvertRecordArray(JArray array, TableFieldSchema fieldSchema)
-            {
-                var ret = new Dictionary<string, object>[array.Count];
-                for (int i = 0; i < ret.Length; i++)
-                {
-                    JObject value = (JObject)array[i];
-                    ret[i] = ConvertRecord((JObject)value["v"], fieldSchema);
-                }
-                return ret;
-            }
-
-            private static Dictionary<string, object> ConvertRecord(JObject record, TableFieldSchema fieldSchema)
-            {
-                var fields = fieldSchema.Fields;
-                JArray values = (JArray) record["f"];
-                if (values.Count != fields.Count)
-                {
-                    throw new InvalidOperationException($"Record had {values.Count} entries; expected {fields.Count}");
-                }
-                var ret = new Dictionary<string, object>(fields.Count);
-                for (int i = 0; i < fields.Count; i++)
-                {
-                    var field = fields[i];                    
-                    var token = values[i]["v"];
-                    Console.WriteLine($"Converting field {field.Name} from token type {token.Type} - token {token}");
-                    ret[field.Name] = ConvertSingleValue(token.Type == JTokenType.String ? (string)token : (object)token, field);
-                }
-                return ret;
-            }
         }
     }
 }
